Updated parse4cn1 push test app
