+++ test/CN1TestApp/nbproject/project.properties
@@ -4,7 +4,8 @@ annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
 application.title=parse4cn1TestApp
-application.vendor=Codename One
+application.vendor=SMash ICT Solutions
+auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -37,7 +38,9 @@ file.reference.impl-stubs=lib/impl/stubs
 file.reference.JavaSE.jar=JavaSE.jar
 file.reference.test-resources=test/resources
 includes=**
+jar.archive.disabled=${jnlp.enabled}
 jar.compress=true
+jar.index=${jnlp.enabled}
 javac.classpath=\
     ${file.reference.CLDC11.jar}:\
     ${file.reference.CodenameOne.jar}:\
@@ -71,7 +74,20 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
+jnlp.codebase.type=no.codebase
+jnlp.descriptor=application
+jnlp.enabled=false
+jnlp.mixed.code=default
+jnlp.offline-allowed=false
+jnlp.signed=false
+jnlp.signing=
+jnlp.signing.alias=
+jnlp.signing.keystore=
 main.class=com.codename1.impl.javase.Simulator
+# Optional override of default Codebase manifest attribute, use to prevent RIAs from being repurposed
+manifest.custom.codebase=
+# Optional override of default Permissions manifest attribute (supported values: sandbox, all-permissions)
+manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
